Move `HttpRequestExecutor` to `firebase-web` module
